* switch to short[]

+++ karaoke.core/src/main/java/com/damn/karaoke/core/controller/processing/IToneDetector.java
@@ -1,5 +1,5 @@
 package com.damn.karaoke.core.controller.processing;
 
 public interface IToneDetector {
-    int analyze(byte[] data, int read);
+    int analyze(short[] data, int read);
 }
